Fixes #114
