+++ metal/contrib/modules/__init__.py
@@ -1,0 +1,11 @@
+from .lstm_module import LSTMModule, Encoder, EmbeddingsEncoder
+from .resnet_cifar10 import ResNetModule
+from .sparse_linear_module import SparseLinearModule
+
+__all__ = [
+    "LSTMModule",
+    "Encoder",
+    "EmbeddingsEncoder",
+    "ResNetModule",
+    "SparseLinearModule",
+]
