--- metal/modules/__init__.py
@@ -1,12 +1,0 @@
-from .identity_module import IdentityModule
-from .lstm_module import LSTMModule, Encoder, EmbeddingsEncoder, CNNEncoder
-from .sparse_linear_module import SparseLinearModule
-
-__all__ = [
-    "IdentityModule",
-    "LSTMModule",
-    "Encoder",
-    "EmbeddingsEncoder",
-    "CNNEncoder",
-    "SparseLinearModule",
-]
